+++ e2e/setup/index.ts
@@ -1,6 +1,12 @@
-// call first
-import './setupSelectors';
-// call before using fabric
-import './setupCoverage';
-// call at the end - navigates the page
-import './setupApp';
+import setupApp from './setupApp';
+import setupCoverage from './setupCoverage';
+import setupSelectors from './setupSelectors';
+
+export default () => {
+  // call first
+  setupSelectors();
+  // call before using fabric
+  setupCoverage();
+  // call at the end - navigates the page
+  setupApp();
+};